+++ pkg/resource/logging-config/alloy-logging-config.go
@@ -23,6 +23,7 @@ var (
 	alloyLoggingConfigTemplate *template.Template
 
 	supportPodLogs = semver.MustParse("1.7.0")
+	supportVPA     = supportPodLogs
 )
 
 func init() {
@@ -46,7 +47,7 @@ func GenerateAlloyLoggingConfig(lc loggedcluster.Interface, observabilityBundleV
 		IsWorkloadCluster                bool
 		SecretName                       string
 		SupportPodLogs                   bool
-		LogsEvents                       bool
+		SupportVPA                       bool
 	}{
 		AlloyConfig:                      alloyConfig,
 		DefaultWorkloadClusterNamespaces: defaultNamespaces,
@@ -54,7 +55,8 @@ func GenerateAlloyLoggingConfig(lc loggedcluster.Interface, observabilityBundleV
 		SecretName:                       common.AlloyLogAgentAppName,
 		// Observability bundle in older versions do not support PodLogs
 		SupportPodLogs: observabilityBundleVersion.GE(supportPodLogs),
-		LogsEvents:     true,
+		// Observability bundle in older versions do not support VPA
+		SupportVPA: observabilityBundleVersion.GE(supportVPA),
 	}
 
 	err = alloyLoggingConfigTemplate.Execute(&values, data)
@@ -80,6 +82,7 @@ func generateAlloyConfig(lc loggedcluster.Interface, observabilityBundleVersion 
 		BasicAuthUsernameEnvVarName string
 		BasicAuthPasswordEnvVarName string
 		SupportPodLogs              bool
+		InsecureSkipVerify          bool
 	}{
 		ClusterID:                   clusterName,
 		Installation:                lc.GetInstallationName(),
@@ -90,7 +93,8 @@ func generateAlloyConfig(lc loggedcluster.Interface, observabilityBundleVersion 
 		BasicAuthUsernameEnvVarName: loggingsecret.AlloyBasicAuthUsernameEnvVarName,
 		BasicAuthPasswordEnvVarName: loggingsecret.AlloyBasicAuthPasswordEnvVarName,
 		// Observability bundle in older versions do not support PodLogs
-		SupportPodLogs: observabilityBundleVersion.GE(supportPodLogs),
+		SupportPodLogs:     observabilityBundleVersion.GE(supportPodLogs),
+		InsecureSkipVerify: lc.IsInsecureCA(),
 	}
 
 	err := alloyLoggingTemplate.Execute(&values, data)