+++ CHANGELOG.md
@@ -10,6 +10,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add kubernetes events logging in Alloy.
+- Add support for Private CAs in alloy logs.
+
+### Changed
+
+- Disable grafana agent usage data reporting.
+
+### Fixed
+
+- Fix logging-config unit tests
+
+## [0.15.2] - 2024-11-13
+
+### Added
+
+- Enable VPA on alloy if the deployed alloy version supports it.
 
 ## [0.15.1] - 2024-11-04
 
@@ -379,7 +394,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add '--vintage' toggle
 - Add controller for Workload Management Cluster using cluster.x-k8s.io/v1beta1```
 
-[Unreleased]: https://github.com/giantswarm/logging-operator/compare/v0.15.1...HEAD
+[Unreleased]: https://github.com/giantswarm/logging-operator/compare/v0.15.2...HEAD
+[0.15.2]: https://github.com/giantswarm/logging-operator/compare/v0.15.1...v0.15.2
 [0.15.1]: https://github.com/giantswarm/logging-operator/compare/v0.15.0...v0.15.1
 [0.15.0]: https://github.com/giantswarm/logging-operator/compare/v0.14.0...v0.15.0
 [0.14.0]: https://github.com/giantswarm/logging-operator/compare/v0.13.0...v0.14.0